+++ Sources/KeyboardKit/_Deprecated/Gestures+Deprecated.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public extension Gestures {
+    
+    @available(*, deprecated, renamed: "Keyboard.SpaceDragSensitivity")
+    typealias SpaceDragSensitivity = Keyboard.SpaceDragSensitivity
+    
+    @available(*, deprecated, renamed: "Keyboard.SpaceLongPressBehavior")
+    typealias SpaceLongPressBehavior = Keyboard.SpaceLongPressBehavior
+}