+++ Tests/KeyboardKitTests/Mocks/MockDragGestureHandler.swift
@@ -9,15 +9,6 @@ import CoreGraphics
 import MockingKit
 @testable import KeyboardKit
 
-class MockDragGestureHandler: Mock, DragGestureHandler {
-    
-    lazy var handleDragGestureRef = MockReference(handleDragGesture)
-    
-    func handleDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint) {
-        call(handleDragGestureRef, args: (startLocation, currentLocation))
-    }
-}
-
 class MockSpaceDragGestureHandler: Gestures.SpaceDragGestureHandler, Mockable {
 
     var mock = Mock()